requirements.txt
desc-opt
ipykernel
matplotlib
nbformat
numpy
scipy